Fix image sizing behavior
Previously, there were situations in which the image object was a different size than the content box of the image element, causing the border radius to disappear (since the border radius only applies to the content box). I changed object-fit from "scale-down" to "contain" and set the CSS width to min-content to fix this behavior. I also added an auto inline margin to center images horizontally and set their maximum height to the height of the viewport (in svh to account for mobile screens) minus 3rem to match the padding on the text column.